requirements.txt
streamlit
tensorflow
numpy
pillow
matplotlib
requests